fix context object on path being ignored.
Note that if your binding expression uses context.property it means a context object on the context.

If you want to use the context directly don't define "context." just the property.

But if you MUST for what ever reason use the $context.property that means the property on the actual binding context.